Fixed compilation error on Ubuntu 10.04 (gcc 4.4.3). Without this cast I get this error: ‘%lld’ expects type ‘long long int’, but argument 2 has type ‘LW_INT64’
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@62457 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lsass/tools/lw-lsa/authenticate_user.c
@@ -228,12 +228,12 @@ PrintAuthUserInfo(
     PrintDataBlob(pInfo->pLmSessionKey);
     printf("\tLogonCount: %d\n", pInfo->LogonCount);
     printf("\tBadPasswordCount: %d\n", pInfo->BadPasswordCount);
-    printf("\tLogonTime: %lld\n", pInfo->LogonTime);
-    printf("\tLogoffTime: %lld\n", pInfo->LogoffTime);
-    printf("\tKickoffTime: %lld\n", pInfo->KickoffTime);
-    printf("\tLastPasswordChange: %lld\n", pInfo->LastPasswordChange);
-    printf("\tCanChangePassword: %lld\n", pInfo->CanChangePassword);
-    printf("\tMustChangePassword: %lld\n", pInfo->MustChangePassword);
+    printf("\tLogonTime: %lld\n", (long long int) pInfo->LogonTime);
+    printf("\tLogoffTime: %lld\n", (long long int) pInfo->LogoffTime);
+    printf("\tKickoffTime: %lld\n", (long long int) pInfo->KickoffTime);
+    printf("\tLastPasswordChange: %lld\n", (long long int) pInfo->LastPasswordChange);
+    printf("\tCanChangePassword: %lld\n", (long long int) pInfo->CanChangePassword);
+    printf("\tMustChangePassword: %lld\n", (long long int) pInfo->MustChangePassword);
     printf("\tLogonServer: %s\n", LW_RTL_LOG_SAFE_STRING(pInfo->pszLogonServer));
     printf("\tLogonScript: %s\n", LW_RTL_LOG_SAFE_STRING(pInfo->pszLogonScript));
     printf("\tProfilePath: %s\n", LW_RTL_LOG_SAFE_STRING(pInfo->pszProfilePath));
